SCB-338 extract AccessLogGenerator
